fix erroneous component setup
